feat(add-ask-cmd): move check_ask into the default AskCommand implementation
