+++ cargo-projects/vivs/src/commands.rs
@@ -6,12 +6,15 @@ use self::ttl::TTL_CMD;
 use crate::data_chunk::DataChunkError;
 use crate::parser::Parser;
 use crate::utils::{unknown_cmd_err, NO_CMD_ERR};
-use crate::{Connection, DataStore, GenericResult};
+use crate::{ClusterInstanceConfig, Connection, DataStore, GenericResult};
 use asking::Ask;
+use core::str;
 use delete::Delete;
 use get::Get;
 use ping::Ping;
 use set::Set;
+use std::env::current_dir;
+use tokio::fs;
 use ttl::Ttl;
 
 pub mod asking;
@@ -31,14 +34,6 @@ pub enum Command {
     Ask(Ask),
     Unknown(String),
     None,
-}
-
-impl Default for Command {
-    fn default() -> Self {
-        Command::Ask(Ask {
-            command: Box::new(Command::Get(Get { key: None })),
-        })
-    }
 }
 
 pub enum DataType {
@@ -73,10 +68,44 @@ impl From<&str> for ParseCommandErr {
 }
 
 pub trait AskCommand {
-    fn check_ask(
-        &self,
-        conn: &mut Connection,
-    ) -> impl std::future::Future<Output = Option<(u16, String)>>;
+    async fn check_ask(&self, key: &str, conn: &mut Connection) -> Option<(u16, String)> {
+        // check if config exists, we'll most likely need to store it in memory to avoid constant IO (?)
+        let own_addr = conn.own_addr().unwrap().to_string();
+
+        let port = own_addr.split(":").nth(1).unwrap_or_default();
+        let current_dir = current_dir().unwrap();
+        let node_config = format!("{}/{}.toml", current_dir.display(), port);
+
+        // When no cluster (<port>.conf) file is found,
+        // We can assume that Vivs is not running in the cluster mode.
+        // Therefore normal processing of incoming command should take place.
+        let Ok(file_contents) = fs::read(node_config).await else {
+            return None;
+        };
+
+        let file_contents = str::from_utf8(&file_contents[0..]).unwrap();
+        let nodes = toml::from_str::<ClusterInstanceConfig>(file_contents).unwrap();
+
+        // Work out a cell / hash slot
+        const X25: crc::Crc<u16> = crc::Crc::<u16>::new(&crc::CRC_16_IBM_SDLC);
+        let key_hash = X25.checksum(key.as_bytes()) % 16384;
+
+        // Iterate over all current nodes in the cluster
+        for (ip, config) in nodes {
+            let cell_range = config.position.0..config.position.1;
+
+            let is_in_range = cell_range.contains(&key_hash.into());
+            if own_addr == ip && is_in_range {
+                return None;
+            }
+
+            if is_in_range {
+                return Some((key_hash, ip));
+            }
+        }
+
+        None
+    }
 }
 
 pub trait CommonCommand {
